requirements.txt
streamlit
pandas
icalendar
python-docx